requirements.txt
Cython==0.29.32
boto3==1.36.24
botocore==1.36.24
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
gensim==4.3.3
medspacy==1.3.1
medspacy_quickumls==3.2
nltk==3.7
numba==0.58.1
numpy==1.24.3
pandas==2.0.0
pickleshare==0.7.5
plotly==6.0.0
PyFastNER==1.0.8
PyRuSH>=1.0.8
pysbd==0.3.4
quickumls==1.4.2
regex==2022.3.2
spacy==3.5.4
spacy-legacy==3.0.11
srsly==2.4.3
thinc==8.1.10
tqdm==4.63.0
umap-learn==0.5.7
unqlite==0.9.3
wasabi==0.9.1
wcwidth==0.2.5
zipp==3.7.0
sentence-transformers==3.2.1
faiss-cpu==1.7.4
torch>=1.13.0
hdbscan==0.8.33
scikit-learn==1.3.2
fastapi==0.110.2
uvicorn[standard]==0.25.0
python-multipart==0.0.9